move metadata helper
