+++ provider/pkg/metadata/overrides.go
@@ -15,10 +15,18 @@
 package metadata
 
 import (
+	"context"
 	"strconv"
 	"strings"
 	"time"
 
+	"github.com/pulumi/pulumi-kubernetes/provider/v4/pkg/await/condition"
+	"github.com/pulumi/pulumi-kubernetes/provider/v4/pkg/logging"
+	appsv1 "k8s.io/api/apps/v1"
+	appsv1beta1 "k8s.io/api/apps/v1beta1"
+	appsv1beta2 "k8s.io/api/apps/v1beta2"
+	batchv1 "k8s.io/api/batch/v1"
+	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 )
@@ -63,4 +71,53 @@ func DeletionPropagation(obj *unstructured.Unstructured) metav1.DeletionPropagat
 	default:
 		return metav1.DeletePropagationForeground
 	}
+}
+
+// DeletedCondition inspects the object's annotations and returns a
+// condition.Satisfier appropriate for using when awaiting deletion.
+//
+// The "inputs" parameter is the source of truth for user-provided annotations,
+// but it is not guaranteed to be named. The "obj" parameter should be used for
+// conditions.
+func DeletedCondition(
+	ctx context.Context,
+	source condition.Source,
+	clientset clientGetter,
+	logger *logging.DedupLogger,
+	inputs *unstructured.Unstructured,
+	obj *unstructured.Unstructured,
+) (condition.Satisfier, error) {
+	if IsAnnotationTrue(inputs, AnnotationSkipAwait) && allowsSkipAwaitWithDelete(inputs) {
+		return condition.NewImmediate(logger, obj), nil
+	}
+	getter, err := clientset.ResourceClientForObject(obj)
+	if err != nil {
+		return nil, err
+	}
+	return condition.NewDeleted(ctx, source, getter, logger, obj)
+}
+
+// allowsSkipDelete returns true for legacy types which support buggy skipAwait
+// behavior during delete.
+// See also:
+// https://github.com/pulumi/pulumi-kubernetes/issues/1232
+// https://github.com/pulumi/pulumi-kubernetes/issues/3154
+func allowsSkipAwaitWithDelete(inputs *unstructured.Unstructured) bool {
+	switch inputs.GroupVersionKind() {
+	case corev1.SchemeGroupVersion.WithKind("Namespace"),
+		corev1.SchemeGroupVersion.WithKind("Pod"),
+		corev1.SchemeGroupVersion.WithKind("ReplicationController"),
+		appsv1.SchemeGroupVersion.WithKind("DaemonSet"),
+		appsv1beta1.SchemeGroupVersion.WithKind("DaemonSet"),
+		appsv1beta2.SchemeGroupVersion.WithKind("DaemonSet"),
+		appsv1.SchemeGroupVersion.WithKind("Deployment"),
+		appsv1beta1.SchemeGroupVersion.WithKind("Deployment"),
+		appsv1beta2.SchemeGroupVersion.WithKind("Deployment"),
+		appsv1.SchemeGroupVersion.WithKind("StatefulSet"),
+		appsv1beta1.SchemeGroupVersion.WithKind("StatefulSet"),
+		appsv1beta2.SchemeGroupVersion.WithKind("StatefulSet"),
+		batchv1.SchemeGroupVersion.WithKind("Job"):
+		return true
+	}
+	return false
 }
